发布文章

+++ leadnews-model/src/main/java/com/heima/model/wemedia/pojos/WmNews.java
@@ -53,7 +53,7 @@ public class WmNews implements Serializable {
      3 多图文章
      */
     @TableField("type")
-    private Short type;
+    private Integer type;
 
     /**
      * 图文频道ID
